+++ plugins/visualization/simple-visualization/src/box-algorithms/ovpCBoxAlgorithmTopographicMap2DDisplay.h
@@ -23,7 +23,7 @@ namespace OpenViBEPlugins
 			uint64_t getClockFrequency() override;
 			bool initialize() override;
 			bool uninitialize() override;
-			bool processInput(const uint32_t index) override;
+			bool processInput(const size_t index) override;
 			bool processClock(OpenViBE::Kernel::IMessageClock& messageClock) override;
 			bool process() override;
